fix(compat): MSVC /W4 /WX portability in phaseY14 shader sources
Six sites in test sources ported Jun 22 (verified on Dawn/clang only) broke the
Windows MSVC /W4 /WX build of `cts`, blocking any current-main Windows run.
Behavior-preserving portability fixes only:

- texture_utils.cpp:6320  C4267 size_t->uint32_t narrowing -> static_cast<uint32_t>
- texture_utils.cpp:6384  C4244 llround(__int64)->double   -> static_cast<double>
- subgroup_util.cpp:249/269  C4996 std::sscanf -> localized #pragma 4996
  (same pattern as common/harness.cpp; sscanf_s is non-portable)
- unary_ranges_common.h:91/92  C4146 unary minus on unsigned -> 0ull - x
  (INT64_MIN-safe, identical two's-complement result)
- cos/sin/tan.spec.cpp  C2065 M_PI undefined -> #define _USE_MATH_DEFINES before <cmath>

Verified: cmake --build build-yawgpu --config Release --target cts -> clean,
cts.exe links; --list catalog intact. No spec/listing/logic changes.

+++ src/webgpu/shader/execution/expression/call/builtin/texture_sampling/texture_utils.cpp
@@ -6317,7 +6317,7 @@ TexelGenerator makeRandomDepthComparisonTexelGenerator(WGPUTextureFormat format,
                                           static_cast<int64_t>(d)});
         const double normalized = clamp01(static_cast<double>(rnd) / 4294967295.0);
         const uint32_t fi = static_cast<uint32_t>(normalized * (kFixedValues.size() - 1));
-        const double value = kFixedValues[std::min<uint32_t>(fi, kFixedValues.size() - 1u)];
+        const double value = kFixedValues[std::min<uint32_t>(fi, static_cast<uint32_t>(kFixedValues.size() - 1u))];
         texel.values[0] = quantizeDepthStencilValue(quantFormat, value);
         return texel;
     };
@@ -6381,7 +6381,7 @@ double ulpFromZero(double value, ComponentDataType type, uint32_t bits) {
             const double scale = type == ComponentDataType::Snorm
                 ? static_cast<double>((1 << (bits - 1)) - 1)
                 : static_cast<double>(bits >= 32 ? 4294967295.0 : ((1u << bits) - 1u));
-            return std::llround(value * scale);
+            return static_cast<double>(std::llround(value * scale));
         }
         case ComponentDataType::Float:
         case ComponentDataType::Ufloat: {
